+++ scarletio/core/traps/tests/test__FutureWrapperAsync.py
@@ -1,12 +1,14 @@
 import vampytest
 
-from ..future_wrapper_async import FutureWrapperAsync
-from ..future import Future
-from ...top_level import get_event_loop, create_event_loop
-from ...time import LOOP_TIME
-from ..task import Task
 from ....utils import is_coroutine_function
+
 from ...event_loop import EventThread
+from ...time import LOOP_TIME
+from ...top_level import create_event_loop, get_event_loop
+
+from ..future import Future
+from ..future_wrapper_async import FutureWrapperAsync
+from ..task import Task
 from ..task_suppression import skip_poll_cycle