+++ scarletio/tools/asynchronous_interactive_console/__init__.py
@@ -1,3 +1,4 @@
+from .auto_completer import *
 from .console import *
 from .console_helpers import *
 from .editors import *
@@ -5,6 +6,7 @@ from .history import *
 
 
 __all__ = (
+    *auto_completer.__all__,
     *console.__all__,
     *console_helpers.__all__,
     *editors.__all__,